requirements.txt
huggingface_hub>=0.20.3
sentence-transformers>=2.5.1
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
beautifulsoup4>=4.12.3
markdown>=3.6
pypdf>=4.0.1
python-docx>=1.1.0
striprtf>=0.0.26
python-pptx>=0.6.23
python-dotenv>=1.0.0
